Extract the MedMeAppInterface
